Improve end turn tool handling
